requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# HTTP 客户端（异步）
httpx==0.26.0

# 日志
loguru==0.7.2

# 其他工具
python-dotenv==1.0.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
